Refactor FeaturedWork component structure to utilize FeaturedWorkClient for improved data fetching and rendering. Update imports and remove obsolete components for cleaner code organization.
