--- components/home/featured-work.tsx
@@ -1,12 +1,0 @@
-import React from "react"
-import FeaturedWorkList from "../featured-work-list"
-
-export default async function FeaturedWork() {
-  return (
-    <div className="mb-10">
-      <h2>See My Featured Work</h2>
-      <div></div>
-      <FeaturedWorkList />
-    </div>
-  )
-}
