+++ package.json
@@ -15,15 +15,15 @@
     "babel-plugin-styled-components": "^1.13.3",
     "emailjs-com": "^2.4.1",
     "framer-motion": "^5.6.0",
-    "next": "^14.0.3",
+    "next": "14.2.33",
     "next-sanity": "^4.3.2",
     "pace-progressbar": "^1.0.9",
     "pg": "^8.7.1",
     "prop-types": "^15.7.2",
-    "react": "^18.2.0",
+    "react": "18",
     "react-alice-carousel": "^1.17.3",
     "react-countup": "^4.3.3",
-    "react-dom": "^18.2.0",
+    "react-dom": "18",
     "react-hamburger-menu": "^1.1.1",
     "react-helmet": "^5.2.1",
     "react-icons": "^3.9.0",
